feat: Get things working
Made changes that allow the did-indy client to work. Unfortunately, it
looks like the driver-did-indy repo is missing a fastapi import when
you're just importing it for the client (they shouldn't be importing it
at all if it's not being used!). With the changes made here, we are now:
- Requesting into the did-indy-driver when the call comes down the
  chain from ACA-Py
- Calling into the did-indy-driver to perform initial setup so that we
  can work with it later

Things may be in a bit of a hackish state right now, but big
breakthroughs have been made that need to be shared with the group. It's
working and we can finally start replacing the bits and pieces.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>