refactor: Author creation moved to context injection
Big refactor in the way we're handling the did-indy client, as well as
fixed all the issues we were having. I may have reverted some of Athen's
changes in terms of API keys and whatnot for the did indy driver,
however, that is only because I was pushing as fast and hard as possible
to get the workflow running. I will work with Athen to restore those
changes and polish things up when we return to work. Lot's of progress
has been made! A couple of hacks remain in the code, a lot of things
need cleaning up, but all-in-all, things are working and that's a
fantastic milestone and a good place to call it for the night.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>